Reference

Contents

Reference

Contents

Index

Index

GlissADe.Cache

GlissADe.Cell

GlissADe.Solution

GlissADe.Solver

GlissADe.Solver

GlissADe.GreenGaussGradient!

GlissADe.MeshQuality

GlissADe.RCM

GlissADe.RCM

GlissADe._DCM

GlissADe._areas

GlissADe._bi_transforms

GlissADe._center

GlissADe._check_valid_face

GlissADe._check_valid_int

GlissADe._cross

GlissADe._dot

GlissADe._edge_centers

GlissADe._edge_lengths

GlissADe._local_coords

GlissADe._mag

GlissADe._mag2

GlissADe._mean_delta

GlissADe._meshcomputations

GlissADe._ncentroid

GlissADe._ncross

GlissADe._neighbours

GlissADe._normalize

GlissADe._normalize!

GlissADe._normals

GlissADe._parsefacelabels_openfoam

GlissADe._parsefaces_openfoam

GlissADe._parsepoints_openfoam

GlissADe._sum_edges

GlissADe._sum_meandelta

GlissADe._surface_grad!

GlissADe.adjMatrix

GlissADe.adjgraph

GlissADe.adjgraph

GlissADe.cellsInsideBoundingPolygon

GlissADe.centralInterpolate!

GlissADe.centralInterpolateParams!

GlissADe.checkDry

GlissADe.computeAverage

GlissADe.computeChunkAverage

GlissADe.computeDegrees

GlissADe.computeFlux

GlissADe.computePressureResidual

GlissADe.computeResidual

GlissADe.computeTimeStep

GlissADe.compute_chunk_edge_velocity

GlissADe.convertToMatrix

GlissADe.findRegularPolygon

GlissADe.gamma!

GlissADe.gammaParams!

GlissADe.generateLinearSystem

GlissADe.getIds!

GlissADe.implicit_solve

GlissADe.init

GlissADe.initWriter

GlissADe.initializeGeometry

GlissADe.linearInterpolate!

GlissADe.meshbounds

GlissADe.meshbounds

GlissADe.muIDefault

GlissADe.parsemesh

GlissADe.plotmesh

GlissADe.preassembleLinearSystem

GlissADe.preprocess

GlissADe.relax!

GlissADe.relaxMomentum!

GlissADe.relaxThickness!

GlissADe.repack

GlissADe.resetCells

GlissADe.saveAt

GlissADe.saveSolution

GlissADe.scalingFactor

GlissADe.solve

GlissADe.solveLinearSystem

GlissADe.sum_dry

GlissADe.testInside

GlissADe.totalDryCells

GlissADe.updateMomentum!

GlissADe.updateSol!

GlissADe.updateThickness!

GlissADe.upwindInterpolate!

GlissADe.upwindParams!

GlissADe.writeFileToVTK

GlissADe.writeToVTK

GlissADe.@maybe_threads

GlissADe.Cache — Type

            Cache{T,S,W}

cache struct for in-place interpolation on each thread used for parallel execution of the implicit solver. This structure is of internal use and is not expected to be used outside of the library.

DataTypes

T - Datatype used to store geometry. Should be of type Dual when differentiating with geometry.

S - Datatype used for indexing. Should be of an integer type.

W - Datatype used for variables. Should be Dual when differentiation is performed.

Fields

ids::Vector{S} - Ids to be looked at for interpolation. ids[1] will always be the owner cell index and the other will be the neighbour.

If ids[2] == ids[1] then the edge is part of boundary. Has a fixed size 2.

params_central::Vector{T} - Used for storing parameters when performing linear/central interpolation. Has a fixed size 2.

params_upwind::Vector{W} - Used for storing parameters when performing first-order upwind interpolation. Has a fixed size 2.

params_gamma::Vector{W} - Used for storing parameters when performing the hybrid gamma interpolation. Has a fixed size 2.

vars_sca::Vector{W} - Used for storing scalar variables when performing interpolation. Has a fixed size 2.

vars_vec::Vector{Vector{W}} - Used for storing 3D vector variables when performing interpolation. Has a fixed size 2.

sca_e::Vector{W} - Used for storing the results of a scalar interpolation. Has a fixed size of 1 scalar.

vec_e::Vector{W} - Used for storing the results of a vector interpolation. Has a fixed size of 1 3D vector.

Iₛ::Matrix{T} - Storage to store curvature transformation. Equals to I - nₚ⊗nₚ

coupling::Matrix{T} - Stores transformations to couple owner and neighbour cells.

vel_i::Vector{W} - Stores the velocity at owner cell.

vel_n::Vector{W} - Stores the velocity at neighbour cell.

GlissADe.Cell — Type

mutable struct Cell{T,S,W}

Structure used for storing geometrical data.

DataTypes

T - Type for numbers for geometrical data. Should be Dual when differentiating with geometry.

S - Type for integers. Defaults to INT_TYPE

W - Type for numbers for state variables. Should be Dual when differentiation is performed.

Fields

idx::S - Face Index

center::Vector{T} - Coordinates of face centroid

vertices::Vector{Vector{T}} - Coordinates of the vertices forming the face.

edge_centers::Vector{Vector{T}} - Coordinates of edge centers of all edges of a face

edge_lengths::Vector{T} - Edge lengths of all edges of a face

normal::Vector{T} - Surface normal at the centroid of a face

area::T - Area of a face

edge_binormals::Vector{Vector{T}} - Binormals orthogonal to edges and pointing outwards for each edge of a face

transform::Vector{Matrix{T}} - Direction Cosine Matrix for transforming variable of current face for each edge.

transform2::Vector{Matrix{T}} - Direction Cosine Matrix for transforming variable of neighbouring face for each edge.

neighbours::Vector{S} - List of neighbouring faces.

h::W - Thickness at this cell

vel::Vector{W} - Velocity (in global coords) in this cell

pb::W - Basal Pressure at this cell

GlissADe.Solution — Type

struct Solution

A Structure containing information to for solution process

Arguments

alpha - Pressure Depth Averaged Coefficient. Default is 0.5

zeta - Velocity Depth Average Coefficient. Default is 1.0

rho - Material Density $(kg/m^3)$. Default is 1500.0

basal_stress - Function to compute Momentum Implicit Coefficients for basal stress.

alpha_p - Under relaxation for pressure. Default is 0.5

alpha_u - Under relaxation for velocity. Default is 0.5

alpha_h - Under relaxation for thickness. Default is 0.5

MIN_ITERS - Minimum corrections per timestep

MAX_ITERS - Maximum corrections per timestep

pMAXRESIDUAL - Maximum allowed residual for pressure to exit loop

hMAXRESIDUAL - Maximum allowed residual for thickness to exit loop

uMAXRESIDUAL - Maximum allowed residual for velocity to exit loop

hclip - Clip the thickness to 0 if the value is below `hclip`

h_min - Minimum thickness to consider a face wet.

Cells - Given by preprocess

points - Coordinates of the vertices of a mesh.

faces - Connectivities of the vertices of a mesh.

location::String - Location to store the solution in VTK format.

GlissADe.Solver — Type

mutable struct Solver{T,S,W}

DataTypes

T - Storage format for numbers related to geometry. Should be Dual when differentiating with geometry

S - Storage format used for integers.

W - Storage format for state variables. Should be Dual when differentiation is performed.

Fields

alpha_p - Under relaxation for pressure. Default is 0.5

alpha_u - Under relaxation for velocity. Default is 0.5

alpha_h - Under relaxation for thickness. Default is 0.5

MIN_ITERS - Minimum corrections per timestep. Default is 6

MAX_ITERS - Maximum corrections per timestep. Default is 15

pMAXRESIDUAL - Maximum allowed residual for pressure to exit loop. Default is 1e-4

hMAXRESIDUAL - Maximum allowed residual for thickness to exit loop. Default is 1e-4

uMAXRESIDUAL - Maximum allowed residual for velocity to exit loop. Default is 1e-4

hclip - Clip the thickness to 0 if the value is below `hclip. Default is0.0`

h_min - Minimum thickness to consider a face wet. Default is 1e-3

Cells - Given by preprocess

points - Coordinates of the vertices of a mesh.

faces - Connectivities of the vertices of a mesh.

location::String - Location to store the solution

GlissADe.Solver — Method

Solver(solution::Solution)

Create Solver object and fill with default values if solution fields left uninitialized.

GlissADe.GreenGaussGradient! — Method

GreenGaussGradient!(cache, Cells, idx, var, vars)

Compute the gradient using cell-based green gauss gradient scheme.

GlissADe.MeshQuality — Method

MeshQuality(Cells; threads=true)

Compute the orthogonal quality of the mesh.

GlissADe.RCM — Method

RCM(adjgr, degrees)

Reverse Cuthill-McKee node-renumbering algorithm. See also: adjgraph, computeDegrees

Arguments

adjgr - Adjacency List of the matrix. Computed using adjgraph

degrees - Degrees of each node in the graph. Computed using computeDegrees

GlissADe.RCM — Method

RCM(A::SparseMatrixCSC; sortbydeg = true)

Reverse Cuthill-McKee node-renumbering algorithm.

Compute the adjacency graph from a sparse matrix. The sparse matrix A is assumed to be symmetric. The results will be wrong if it isn't.

sortbydeg: Should the neighbor lists be sorted by column degree? The default is true, but often results of very similar quality are obtained when this is set to false and the lists are not sorted. The second option can be much faster, as the sorting is expensive when the neighbor lists are long.

GlissADe._DCM — Method

_DCM(local_coords::Vector{Vector{W}})

Computes the Direction Cosine Matrix from local coords to global coords. Equivalent to converting the Vector{Vector} localcoords to Matrix

Arguments

local_coords::Vector{Vector{Real}} - Local coordinate system of a face

GlissADe._areas — Method

_areas(points::Vector{Vector{W}}, faces::Vector{Vector{S}}, centers::Vector{Vector{W}}) where {W<:Real, S<:Integer}

Returns the areas of all faces of a mesh using the vertices points and the centroids centers.

Arguments

points - Coordinates of all vertices of a mesh

faces - Connectivities of all vertices of a mesh

centers - Face centroids of all vertices of a mesh

GlissADe._bi_transforms — Method

_bi_transforms(centers::Vector{Vector{W}}, edge_centers::Vector{Vector{W}}, normals::Vector{Vector{W}}, points::Vector{Vector{W}}, faces::Vector{Vector{S}}, neighbours::Vector{Vector{S}}) where {W<:Real, S<:Integer}

Returns the edge binormals and transformation matrices for all faces of a mesh.

Arguments

centers - Face centroids of all faces of a mesh

edge_centers - Edge centers of all edges of all faces of a mesh

normals - Surface normal at face centroids of all faces of a mesh

points - Coordinates of all vertices of a mesh

faces - Connectivity list of all vertices of a mesh

neighbours - Adjacency list (neighbour list) of all faces of a mesh

GlissADe._center — Method

_center(points::Vector{Vector{W}}, face::Vector{S}) where {W<:Real, S<:Integer}

Computes the centroid (arithmetic mean) given the vertices and connectivity of a face.

Arguments

points - Coordinates of all vertices of a mesh

face - Connectivity of a "face"

GlissADe._check_valid_face — Method

        _check_valid_face(line)

Checks if the given string line represents a valid face.

Arguments

line::String - line which needs to be checked

GlissADe._check_valid_int — Method

        _check_valid_int(s)

Checks if the given variable is of type-constant INT_TYPE.

Arguments

s::AnyDataType - Any Variable

GlissADe._cross — Method

_cross(x::Vector{W},y::Vector{W}) where W<:Real

Returns the cross product of 3D vectors x and y.

GlissADe._dot — Method

_dot(x::Vector{W},y::Vector{W}) where W<:Real

Computes the dot product (inner product) of 3D vectors x and y.

GlissADe._edge_centers — Method

_edge_centers(points::Vector{Vector{W}}, faces::Vector{Vector{S}}) where {W<:Real, S<:Integer}

Returns the edge centers for all edges of all faces of a mesh.

Arguments

points - Coordinates of all vertices of a mesh

faces - Connecitivity list of vertices of a mesh

GlissADe._edge_lengths — Method

_edge_lengths(points::Vector{Vector{W}}, faces::Vector{Vector{S}}) where {W<:Real, S<:Integer}

Returns the edge lengths of all edges of all faces of a mesh.

Arguments

points - Coordinates of all vertices of a mesh

faces - Connectivities of all vertices of a mesh

GlissADe._local_coords — Method

_local_coords(n::Vector{W}, p1::Vector{W}, p2::Vector{W}; axis=2) where W<:Real

Computes the local coordinate system given two vertices p1 and p2 and the normal n.

Arguments

axis = 2 → The 3D vector p2-p1 is taken to be the y axis of the local coordinate system

axis = 1 → The 3D vector p2-p1 is taken to be the x axis of the local coordinate system

GlissADe._mag — Method

_mag(x::Vector{W}) where W<:Real

Computes the Euclidean norm (2-norm) of 3D vector x.

GlissADe._mag2 — Method

_mag2(x::Vector{W},y::Vector{W}) where W<:Real

Computes the Euclidean norm (2-norm) of 3D vector x-y.

GlissADe._mean_delta — Method

computeMeanDelta(Cells)

Computes mean spacing between face centroids. Uses multiple threads if threads=true.

GlissADe._meshcomputations — Method

_meshcomputations(points, faces, neighbours, mesh)

Precompute geometrical information of a parsed mesh and store it in the Cell structure

points - Array containing vertices of the mesh

faces - Array containing connectivities/faces of the mesh

neighbours - Array containing neighbours of each face of the mesh

```

GlissADe._ncentroid — Method

_ncentroid(v1::Vector{W}, v2::Vector{W}, v3::Vector{W}) where W<:Real

Returns the normal to the plane containing 3D vectors v1, v2 and v3 centered at v2.

GlissADe._ncross — Method

_ncross(x::Vector{W},y::Vector{W}) where W<:Real

Expands to _normalize(_cross(x,y)). Returns the normalized cross product of 3D vectors x and y.

GlissADe._neighbours — Method

_neighbours(l, faces)

Computes the neighbours of all faces in the mesh. This algorithm has a time complexity of $O(PE)$ where $P, E$ are the total number of points and edges in the mesh respectively.

Arguments

l = $P$, Total number of points in the mesh

faces - Vector containing information of all the faces in the mesh

GlissADe._normalize! — Method

_normalize!(x::Vector{W}) where W<:Real

Normalizes 3D vector x inplace using its Euclidean norm (2-norm).

GlissADe._normalize — Method

_normalize(x::Vector{W}) where W<:Real

Return 3D vector x normalized using its Euclidean norm (2-norm).

GlissADe._normals — Method

_normals(points::Vector{Vector{W}}, faces::Vector{Vector{S}}) where {W<:Real, S<:Integer}

Returns the face centroid and the surface normal at face centroid.

Arguments

points - Coordinates of the vertices of a mesh

faces - Connecitivity list of the vertices of a mesh

GlissADe._parsefacelabels_openfoam — Method

_parsefacelabels_openfoam(location_faceLabels, correction_factor)

Parse indices of the faces forming the terrain in a text file with OpenFoam format and stores it in INT_TYPE

Arguments

location_faceLabels::String - location of the facelabels file.

correction_factor::INT_TYPE - remove incorrect entries

GlissADe._parsefaces_openfoam — Method

_parsefaces_openfoam(location_faces)

Parse the connectivity list of a mesh in an OpenFOAM format and store it in type INT_TYPE

Arguments

location_faces::String - Location of the "faces" file generated by OpenFOAM avalanche module

GlissADe._parsepoints_openfoam — Method

        _parsepoints_openfoam(location_points)

Parse the points file generated in an OpenFoam format and stores the coordinates in FLOAT_TYPE format.

Arguments

location_points::String - Location of the "points" file generated by OpenFOAM avalanche module

GlissADe._sum_edges — Method

Multithreading auxillary for computeMeanDelta

GlissADe._sum_meandelta — Method

Multithreading auxillary for computeMeanDelta

GlissADe._surface_grad! — Method

_surface_grad!(surface_grad, n)

Updates the matrix surface_grad inplace using the surface normal n.

GlissADe.adjMatrix — Method

    adjMatrix(neighbours)

Computes the Adjacency Matrix of the Mesh given the neighbours

Arguments

neighbours::Vector{Vector{INT_TYPE}} - List of neighbours for each face.

GlissADe.adjgraph — Method

adjgraph(conn, nfens)

Compute the adjacency graph from the array of connectivities of the elements in the mesh.

Example

conn = [9 1 8 4;
       1 3 2 8;
       8 2 7 5;
       2 6 7 7];
nfens = 9;
adjgraph(conn, nfens)

should produce 9-element Array{Array{Int64,1},1}: [9, 8, 4, 3, 2] [1, 3, 8, 7, 5, 6] [1, 2, 8] [9, 1, 8] [8, 2, 7] [2, 7] [8, 2, 5, 6] [9, 1, 4, 3, 2, 7, 5] [1, 8, 4]

GlissADe.adjgraph — Method

adjgraph(A; sortbydeg)

Compute the adjacency graph from a sparse matrix.

The sparse matrix A is assumed to be symmetric. The results will be wrong if it isn't.

sortbydeg: Should the neighbor lists be sorted by column degree? The default is true, but often results of very similar quality are obtained when this is set to false and the lists are not sorted. The second option is much faster, as the sorting is expensive.

GlissADe.cellsInsideBoundingPolygon — Method

cellsInsideBoundingPolygon(polygon, Cells)

Returns a list of indices of faces lying inside a polygon given by polygon.

Arguments

polygon::Vector{Vector{number}} - A list of vertices in order connecting to form a polygon Cells::Vector{Cell{T,S,W}} - Generated by preprocess

GlissADe.centralInterpolate! — Method

centralInterpolate!(Cells, Cell_idx, edge_idx, cache; IDS_PRECOMPUTED, PARAMS_PRECOMPUTED, scalar)

Performs central interpolation on variables stored in cache.vars_vec if scalar=false or cache.vars_sca if scalar=true

Arguments

Cells::Vector{Cell{T,S,W}} - Given by preprocess

Cell_idx - Index of Current Face

Edge_idx - Edge Index

cache::Cache{T,S,W} - Cache generated for edge interpolation

IDS_PRECOMPUTED - Skip computation of ids if set true

PARAMS_PRECOMPUTED - Skip computation of parameters if set true

scalar - Use cache.vars_sca instead of cache.vars_vec if set true

GlissADe.centralInterpolateParams! — Method

centralInterpolateParams!(params, Cells, Cell_idx, edge_idx, ids)

Computes the parameters for central interpolation between edges.

Arguments

params - storage for two values of type FLOAT_TYPE or Dual

Cells::Vector{Cell{T,S,W}} - Given by preprocess

Cell_idx - Index of current cell

edge_idx - Edge Index

GlissADe.checkDry — Method

checkDry(solver, ids, idx)

Check if the face at index idx and it's neighbours are dry.

GlissADe.computeAverage — Method

computeAverage(u)

Compute the arithmetic mean of vector u

GlissADe.computeChunkAverage — Method

Multithreading auxillary for computeAverage

GlissADe.computeDegrees — Method

computeDegrees(adjgr)

Compute the degrees of the nodes in the adjacency graph.

Arguments

adjgr - Adjacency graph of a matrix. Computed using adjgraph

Example

conn = [9 1 8 4; 1 3 2 8; 8 2 7 5; 2 6 7 7]; nfens = 9; adjgr = adjgraph(conn, nfens) computeDegrees(adjgr)

julia> degrees = computeDegrees(adjgr) 9-element Array{Int64,1}: 5 6 3 3 3 2 4 7 3

GlissADe.computeFlux — Method

Expands to dot(mₑ, vel_edge).

GlissADe.computePressureResidual — Method

computePressureResidual(Cells, h, pb, vel, caches; threads=true)

Compute the pressure residual after an iteration. Used mainly for implicit solvers.

GlissADe.computeResidual — Method

computeResidual(solver, u, u_avg, A, B)

Computes the residual of the linear system $Au=B$. See docs for more information about residual calculation. INTERNAL

Arguments

solver - Solver struct defining the solution process

u - Current Solution

u_avg - Average Value of the solution

A - Coefficient Matrix of the linear system

B - Source of the linear system

GlissADe.computeTimeStep — Method

computeTimeStep(solver, Cₘ, Δₑ, caches)

Compute the timestep adaptively based on the CFL criterion. $dt = Cₘ*Δₑ/cₑ$ INTERNAL

GlissADe.compute_chunk_edge_velocity — Method

Multithreading auxillary for computeTimeStep

GlissADe.convertToMatrix — Method

convertToMatrix(a)

Converts a vector vector to a matrix of float type. The output matrix will be $N$x$M$ where N = ndims or ncols and M = nrows in the original Vector of Vector a

julia> convertToMatrix([[1.0, 2.0], [3.0, 4.0], [5.0, 6.0]])
2×3 Matrix{Float64}:
 1.0  3.0  5.0
 2.0  4.0  6.0

GlissADe.findRegularPolygon — Method

findRegularPolygon(coord_limits; npoints=4, epsilon=1e-1)

Returns a regular polygon within the given constraints: [xmin, xmax, ymin, ymax].

julia> init(stats = false, threads = false, plots = false);

julia> polygon = findRegularPolygon([0.0, 10.0, 0.0, 10.0], npoints = 4);

julia> [round.(p, digits = 4) for p in polygon]
4-element Vector{Vector{Float64}}:
 [10.1, 5.0]
 [5.0, -0.1]
 [-0.1, 5.0]
 [5.0, 10.1]

GlissADe.gamma! — Method

gamma!(cache, Cells, idx, edge_idx, flux_edge, var, vars; βₘ = 0.5, IDS_PRECOMPUTED=false)

Perform gamma interpolation on the var given by var and the total array vars.

GlissADe.gammaParams! — Method

gammaParams!(cache, Cells, idx, edge_idx, flux_edge, var, vars; βₘ = 0.5)

Compute the parameters for the gamma interpolation scheme.

GlissADe.generateLinearSystem — Method

Pre-allocate the memory for storing linear systems.

GlissADe.getIds! — Method

getIds!(ids, Cells, Cell_idx, edge_idx)

Returns the index of the cells sharing the edge given by Cells[i].edge[edge_idx]

Arguments

ids - Arrays providing storage for 2 integers of type INT_TYPE

Cells::Vector{Cell{T,S,W}} - Given by preprocess

Cell_idx - Index of current cell

edge_idx - Edge Index

GlissADe.implicit_solve — Method

implicit_solve(solver, tspan, Cₘ, saveat, rtol)

The solver process which simulates the flow.

Arguments

solver - As generated from Solver

tspan - Flow simulated in the range [tspan[1], tspan[2]]

Cₘ - Maximum courant number allowed. Defaults to 0.9

saveat - Save the solution at an equal intervals of saveat. If set to 0.0, no interpolation will be done and original solution returned.

rtol - Relative tolerance for linear solvers.

GlissADe.init — Method

init(plots, threads, stats, float_type, int_type, implicit)

Initialize the library and start PyPlot backend if plots=true

Arguments

plots::Bool: toggle side-by-side iteration plotting.

threads::Bool: toggle multi-threading

stats::Bool: toggle debug-printing

float_type::DataType: fallback type for floats.

int_type::DataType: fallback type for ints.

implicit::Bool: whether solve uses the implicit solver. Default true.

GlissADe.initWriter — Method

initWriter(location::String, points, faces)

Initialize Writer for VTK files.

GlissADe.initializeGeometry — Method

initializeGeometry(cells_inside_polygon, Cells, rho; h0 = nothing, u0 = nothing)

Initializes the faces in cells_inside_polygon to have a thickness h0 and velocity u0. Pressure is initialized with a value used for flat surfaces.

GlissADe.linearInterpolate! — Method

linearInterpolate!(out, params, vars)

Computes and stores params[1]*vars[1]+params[2]*vars[2] in out.

GlissADe.meshbounds — Method

meshbounds(Cells)

Computes the mesh span on the $xy$ plane of mesh cell centroids. See also: preprocess

Arguments

Cells::Vector{Cell{T,S,W}} - Cell structure generated by preprocess

GlissADe.meshbounds — Method

meshbounds(points)

Computes the mesh span on the $xy$ plane of mesh vertices.

Arguments

points - Coordinates of all vertices of a mesh.

julia> meshbounds([[0.0, 0.0, 0.0], [1.0, 2.0, 0.0], [-1.0, 3.0, 0.0]])
x_min: -1.0, x_max: 1.0
y_min: 0.0, y_max: 3.0

GlissADe.muIDefault — Method

muIDefault(Cells, h, vel, pb, alpha, zeta, rho, idx)

A default basal stress computation based on muI rheology. INTERNAL

GlissADe.parsemesh — Method

parsemesh(location_points, location_faces, location_faceLabels)

Parse the files of a mesh in OpenFoam format.

location_points::String - Location of "points" file generated by OpenFOAM polymesh.

location_faces::String - Location of "faces" file generated by OpenFOAM polymesh.

location_faceLabels::String - Location of "facelabels" file generated by OpenFOAM polymesh.

GlissADe.plotmesh — Function

plotmesh(points, faces; axis = NamedTuple())
plotmesh(Cells::Vector{Cell}; field = nothing, axis = NamedTuple())

Render mesh geometry natively via Meshes.jl and a Makie backend, without any external tool.

Accepts either the raw (points, faces) returned by parsemesh, or a precomputed Vector{Cell} returned by preprocess.

For the Vector{Cell} form, field paints each face with a per-cell scalar color:

a Vector of scalars, one per cell (must satisfy length(field) == length(Cells));

a Symbol selecting a cell field directly: :h (thickness), :pb (basal pressure), :U, :V, :W (global velocity components), or :speed (velocity magnitude);

a Function applied to each Cell to produce its scalar value;

nothing (default) for bare geometry with no coloring.

The initial camera orientation defaults to looking along the mesh's area-weighted average surface normal, rather than Makie's generic isometric default, since a generic angle tends to show terrain-like meshes almost edge-on. Pass axis (e.g. axis = (azimuth = ..., elevation = ...)) to override this default; any keys given there take priority over the computed default.

Requires a Makie backend to be loaded (e.g. using GLMakie, using WGLMakie, or using CairoMakie). The implementation lives in the GlissADeMakieExt package extension and is not available otherwise.

GlissADe.preassembleLinearSystem — Method

preassembleLinearSystem(Cells::Vector{Cell{T,INT_TYPE,W}}) where {T<:ALLOWED_NUMBERS, W<:ALLOWED_NUMBERS}

Generate Sparse matrices and their preconditoners with the sparsity pattern of the mesh. Also allocated additional arrays for solving partial derivatives of linear systems.

Arguments

Cells::Vector{Cell{T,INT_TYPE,W}} - Discretized Geometry

GlissADe.preprocess — Method

preprocess(points, faces; comp_neighbours=true)

Precompute geometrical information and store it in the Cell structure.

points - coordinates of the vertices of the mesh.

faces - connectivity list of the mesh.

W - Datatype to be used for variables. Should be of type Dual if differentiating with geometry

comp_neighbours - Whether to compute neighbours or read from file. Default is true. In case of incompatibility with mesh, neighbours will be recomputed.

GlissADe.relax! — Method

relax!(var_new, var_old, alpha)

Under relax the field stored in var_new with the old field var_old inplace. alpha controls the amount of relaxation and should be in the range $[0,1]$

GlissADe.relaxMomentum! — Method

relaxMomentum!(A,B,alpha,vel0,Cells)

Under-relax the linear system for momentum equation in-place. Equation Under-relaxation.

GlissADe.relaxThickness! — Method

relaxThickness!(A,B,alpha,h0,Cells; threads=true)

Under relax the linear system for the continuity (thickness) equation in-place. Equation Under-relaxation.

GlissADe.repack — Method

Repack the computed values and partial derivatives for solution of a linear system.

GlissADe.resetCells — Method

resetCells(Cells) where {T<:Real, S<:Integer}

Reset discretized geometry's stored information to zero.

GlissADe.saveAt — Method

Save the solution at prescribed timesteps.

GlissADe.saveSolution — Method

saveSolution(location::String, time, time_steps, sol, iter, points_mat, cells)

Save the solution at the time time. See also: writeToVTK

GlissADe.scalingFactor — Method

scalingFactor(solver, p)

Returns the scaling factor for scaling the pressure equation residual. INTERNAL

GlissADe.solve — Method

solve(solver, tspan; Cₘ=0.9, saveat=0.0, rtol=0.01)

Solve the flow equations over the time span tspan.

Arguments

solver::Solver - Solver object created via Solver.

tspan - Time span (t0, t1) to simulate over.

Cₘ - Maximum Courant number used to control adaptive step sizes. Default 0.9.

saveat - Uniform interval at which to save the solution. 0.0 (default) returns every adaptive timestep instead of resampling onto a uniform grid.

rtol - Relative tolerance used by the linear solve. Default 0.01.

Whether the implicit or explicit update is used is set by init's implicit keyword; only the implicit path is currently implemented.

GlissADe.solveLinearSystem — Method

Solves the linear system and computes its partial derivatives

GlissADe.sum_dry — Method

Multithreading auxillary for totalDryCells.

GlissADe.testInside — Method

testInside(coords, bounding_coords)

Tests whether a point with coordinates coords is inside the region given by bounding_coords: [xmin, xmax, ymin, ymax] INTERNAL

GlissADe.totalDryCells — Method

totalDryCells(solver)

Compute the total number of dry cells at any time step. See also: checkDry.

GlissADe.updateMomentum! — Method

updateMomentum!(solver, Av, Bv, precon_v, dt, t, p, vel0, h0, caches; threads=true, prev_solution=nothing, pprev_solution=nothing)

Solves the momentum equation inplace and stores the result in vel0. INTERNAL

GlissADe.updateSol! — Method

Update function to store the solution after a time-step.

GlissADe.updateThickness! — Method

updateThickness!(solver, Ah, Bh, precon_h, dt, t, p, vel, h0, caches; threads=true, prev_solution=nothing, pprev_solution=nothing)

Solves the continuity (thickness) equation inplace and stores the result in h0. INTERNAL

GlissADe.upwindInterpolate! — Method

upwindInterpolate!(cache, flux_edge; PARAMS_PRECOMPUTED, scalar)

Perform upwind interpolation using cache.vars_vec if scalar=false and cache.vars_sca if scalar=true

Arguments

cache::Cache{T,S,W} - Cache for edge interpolation

flux_edge::W - Flux at edge (computed using central interpolation)

PARAMS_PRECOMPUTED - Skip computation of params if set true

scalar - Use cache.vars_vec if set to false, else use cache.vars_sca

GlissADe.upwindParams! — Method

upwindParams!(params, flux_edge)

Compute Upwind Interpolation parameters based on flux at edge.

GlissADe.writeFileToVTK — Method

writeFileToVTK(filename::String, sol, idx, points_mat, cells)

Write a file with a given filename and sol to disk in Paraview's VTK format. See also: writeToVTK

GlissADe.writeToVTK — Method

writeToVTK(location::String, sol, points, faces)

Write a solution to VTK files at the given location.

Arguments

$location::String$ : Destination of the VTK files

$sol$ : Solution from the solver's output

$points$ : Mesh vertices

$faces$ : Mesh connectivities

GlissADe.@maybe_threads — Macro

maybe_threads(flag::Bool, expr)

Execute a for-loop with threads only if flag==true

Example Usage

julia> @maybe_threads (1==2) for i in 1:10
        println("I love Julia!")
       end